requirements.txt
attr==0.3.2
ConfigParser==6.0.0
contextlib2==21.6.0
Cython==3.0.8
dl==0.1.0
docutils==0.20.1
filelock==3.13.1
HTMLParser==0.0.2
Jinja2==3.1.3
keyring==24.3.0
mock==5.1.0
mysql_connector_repackaged==0.3.1
openapi_spec_validator==0.7.1
Pillow==10.2.0
prance==0.21.8.0
protobuf==3.20.3
PyGithub==2.2.0
PyMySQL==1.1.0
pyOpenSSL==24.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
railroad==0.5.0
redis==5.0.1
Sphinx==7.2.6
sqlescapy==1.0.1
thread==0.1.3
tornado==6.3.3
trove_classifiers==2024.1.31
urllib3_secure_extra==0.1.0
Cython==3.0.8
